refactor:도메인별 엔티티 폴더 추가 후 폴더 구조 개선
