+++ src/main/java/com/k_passs/backend/domain/bill/entity/Bill.java
@@ -1,9 +1,9 @@
-package com.k_passs.backend.domain.bill;
+package com.k_passs.backend.domain.bill.entity;
 
 import com.k_passs.backend.domain.model.entity.BaseEntity;
 import com.k_passs.backend.domain.model.enums.BillType;
 import com.k_passs.backend.domain.model.enums.BillStatus;
-import com.k_passs.backend.domain.user.User;
+import com.k_passs.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
