+++ src/main/java/com/k_passs/backend/domain/user/entity/PointHistory.java
@@ -1,4 +1,4 @@
-package com.k_passs.backend.domain.user;
+package com.k_passs.backend.domain.user.entity;
 
 import com.k_passs.backend.domain.model.entity.BaseEntity;
 import jakarta.persistence.*;
