+++ src/main/java/com/k_passs/backend/domain/user/entity/UserChallenge.java
@@ -1,6 +1,6 @@
-package com.k_passs.backend.domain.user;
+package com.k_passs.backend.domain.user.entity;
 
-import com.k_passs.backend.domain.challenge.Challenge;
+import com.k_passs.backend.domain.challenge.entity.Challenge;
 import com.k_passs.backend.domain.model.entity.BaseEntity;
 import com.k_passs.backend.domain.model.enums.ChallengeStatus;
 import jakarta.persistence.*;
